Add the .rng for XML-Grammar-Vered.
This will be a custom and ad-hoc XML grammar used to notate
src/tutorials/bad-elements/index.html.wml and in the future possibly
other documents because the Website Meta Language markup of bad-elements
is getting out of hard, and also makes it harder to contribute to the page
or translate it.

Vered is a transliteration of the Hebrew word for "Rose" as a reference
to "A Rose by any other name", and is also a common Israeli feminine name.
In Hebrew it is writen ???:

http://en.wiktionary.org/wiki/%D7%95%D7%A8%D7%93